unixsock plugin et alii: Allow passing of arbitrary identifiers to the FLUSH command.
Due to a bug in the parsing of the FLUSH command, identifiers which
contained spaces would not be accepted. This patch introduces
src/utils_parse_option.[ch], which strips off an option in the form of
name="value" from the beginning of a buffer (removing leading and
trailing spaces) and returns the values found.

The Collectd::Unixsock module is updated to automatically quote
identifiers when necessary.

Many other commands may be effected by this bug, too. They'll be fixed
soon.