perl plugin: Added "<Plugin>" configuration block.
Similar to the global "<Plugin>" blocks this may be used to configure Perl
plugins. The oconfig_item_t object is converted to a Perl hash which will be
passed on to the registered configuration callback. A configuration
callback is registered using the TYPE_CONFIG identifier.

The Perl representation of the oconfig_item_t object looks like this:

  {
      'key'      => '<key>',
      'values'   => [ '<val1>', '<val2>', ... ],
      'children' => [ { ... }, { ... }, ... ]
  }

Signed-off-by: Sebastian Harl <sh@tokkee.org>
Signed-off-by: Florian Forster <octo@huhu.verplant.org>