src/plugin.[ch]: Added support for "filter" callbacks.
This type of callback is very similar to write callbacks:

  int (*) (const data_set_t *, value_list_t *)

The return value of that callback indicates what further actions are to be
taken. It may be a bitwise OR of any of the following new flags:
FILTER_NOWRITE or FILTER_NOTHRESHOLD_CHECK. If both flags are used (which
equals the flag FILTER_IGNORE), the value list is ignored completely.

If more than one such callback is available, the respective return values are
bitwise OR'ed. If they "sum" up to FILTER_IGNORE, no further callbacks are
executed (lazy evaluation).

This type of callback may be used for the following purposes:

 * rewrite information included in the value list (e.g. the instance
   name)

 * discard certain information on a fine-grained level based on the
   information provided by the data set and value list

Signed-off-by: Sebastian Harl <sh@tokkee.org>
Signed-off-by: Florian Forster <octo@huhu.verplant.org>