Document unused parameters.
In all cases where unused parameters have to exist either because some API
requires it or for more consistency, they now have been documented using
__attribute__((unused)) (which is automatically disabled in case GCC is not
used to prevent problems with compilers not supporting attributes). In a few
other cases, the parameters have been removed.

This allows the use of -Wunused-parameter (which is enabled by -Wextra as
well, when using -Wall).