java plugin: Added the ability to have `Write' callbacks in Java modules.
I think right now having both, a `Read' and a `Write' function, would be
a problem, because the same thread would enter the JVM twice, possibly
detaching itself from the JVM when the write callback is finished, while
it actually still is in the read callback. Adding a `dispatch thread' or
something similar should take care of this problem.

Other than that, converting `data_set_t' and `value_list_t' to their
Java equivalents and back works fine now.