php-collection: Add support for named graph lists as favorites
php-collection: Add support for named graph lists as favorites

php-collection allowed saving and loading a list of graphs from a
cookie.
This patch changes loading/saving to provide support for naming
of favorite graph lists.
When saving a name has to be provided (up to 30 alphanumerical
characters includeing underscore and hyphen.
For loading a named list can be selected from a dropdown list (this list
is refreshed when it gets focus) of available favorites.

Signed-off-by: Florian Forster <octo@huhu.verplant.org>