bind plugin: Actually handle RR cache numbers as `gauge'..
Hi Florian,

Well a working patch was even less trivial as the quick one from
yesterday evening.

Just using a gauge type is not sufficient (only 0 is written to RRD
files), it's also necessary to adjust the bind_xml_stats_handle_view()
as has been done for the memory statistics

Attached is a patch to bind.c (I skipped types.db) which fixes the
cache data collection.

In addition I attached a graph of the cache evolution and memory (for
memory it might be worth giving it an own type as it's not cumulative
at all)

Bruno