mysql plugin: Added support for "slave notifications".
The new boolean config option "SlaveNotifications" may be used to enable
notifications if the replication slave I/O and / or SQL threads are not
running (and the respective OKAY notifications). Since this is somewhat
related to the time offset between master and slave, "time_offset" is used for
the type information stored in the notification.