uptime plugin: Calculate the boot time under Linux.
Hi Florian,

[...]

About the uptime plugin:
New code attached, it should be fine now (different linux behaviour and some
minor changes). As soon as I can I'll send to you a plotting code sample (with
lots of stuff and RPN expressions) to be added in the contributions dir, if you
like, and edit the wiki page.

Regards,
Marco