mysql plugin: Add `lock' statistics.
Hi,

We evaluate collectd for internal use in our global monitoring system. For that
we need to monitor mysql locks, so as collectd doesn't do this I wrote a little
patch. This patch is based on 4.6.2 version and modify only src/mysql.c, hope
you'll find it usefull. If this patch is acceptable please feel free to
integrate it in your next release.

Regards,

--
Rodolphe Quiedeville
www.bearstech.com