src/plugin.c: Initiate the threshold checking again.
Hi,

I have the same problem in my instalation, exactly the same scenario, I found a
possible explanation in the plugin.c module. While in the 4.5 branch exists a
callback to function ut_check_threshold (defined in utils_threshold.c), in the
4.6.2 (and also in 4.6.1, I think), the callback is missing, so threshold
checking never runs.

In fact, a grep -r ut_check_threshold over src dir, only show the definition of
the function in ut_check_threshold module.

I'm not sure if it's really a bug, but when I patched my code, it works fine
for me :) I'm using the 4.6.2 version from tar.gz.

Best regards,
  Andrés

The regression was introduced in 65954d9b.

Signed-off-by: Florian Forster <octo@huhu.verplant.org>