src/utils_threshold.c: Percentage support in thresholds
Hi all!

I attach a patch to add percentage support in thresholds, like this example:

<Threshold>
  <Type df>
   WarningMax 90
   Percentage true
 </Type>
</Threshold>

The percentage option works like collectd-nagios, that is, calculate the
percentage of the value of the first DS over the total. For df plugin,
for example,
calculate the percentage of the "used" DS.

Bugs and suggestions are welcome :)

Enjoy!

Regards,
  Andres