madwifi plugin: Plugin for detailed information from the MadWifi driver.
Hello

After some time i managed to make a new version of Madwifi plugin. The
main change is that it is possible to finely tune the set of monitored
statistics and just the most important statistics are monitored by
default. Also the number of new data types is reduced (by using type
instances).

Signed-off-by: Florian Forster <octo@verplant.org>