src/utils_cache.c: Implement “uc_{get,set,inc}_hits”.
Hi

I've attached a patch to add hit counter to thresholds, that is, each
time when threhsold raised, then an internal hit counter is incremented,
when the value of the counter raise a specific value setted in
configuration, then the notification is generated and counter is reset.
Here are an example of threshold configuration with hit conter:

<Threshold>
  <Plugin load>
    <Type load>
       WarninMax 1
       Hits 3
    </Type>
   </Plugin>
</Threshold>

In this example the notification will be generated when load value is
greater than 1 for 3 intervals.

Here are two patches, the first one adds hit feature to the collectd
cached and the second one use hits in thresholds,

Of course comments and suggestions are welcome :)

Enjoy!

Regards,
  Andres

Signed-off-by: Florian Forster <octo@leeloo.lan.home.verplant.org>