netapp plugin: Refactored volume usage statistics.
Much of this is like disk, wafl and system statistics before. The
“GetVolumeData” has been renamed to “VolumeUsage” and the
“GetDiskUtil” and “GetSnapUtil” options have been changed, too. The
configuration now looks like this:

    <VolumeUsage>
      Capacity "vol0"
      Capacity "vol1"
      IgnoreSelectedCapacity false
      Snapshot "vol1"
      Snapshot "vol3"
      IgnoreSelectedSnapshot false
    </VolumeUsage>

The code now uses to "ignore lists" to check whether capacity and/or
snapshot information should be collected for a volume. This means the
order in which volumes are listed no longer matters and than you can
use such advanced options as selecting volumes via regular expressions.